Mb delete knots

+++ knotitproject/knotitapp/views/knots/details.py
@@ -7,44 +7,50 @@ from ..connection import Connection
 
 
 def get_knot(knot_id):
-   
+    
     return Knot.objects.get(pk=knot_id)
 
 
 @login_required
 def knot_details(request, knot_id):
-    if request.method == 'GET':
-        knot = get_knot(knot_id)
-        template_name = 'knots/details.html'
-        return render(request, template_name, {'knot': knot})
+    current_user = request.user.id
+    knot = get_knot(knot_id)
+    if knot.user_id == current_user:
+        
+        if request.method == 'GET':
+            knot = get_knot(knot_id)
+            template_name = 'knots/details.html'
+            return render(request, template_name, {'knot': knot})
 
-    elif request.method == 'POST':
-        form_data = request.POST
+        elif request.method == 'POST':
+            form_data = request.POST
 
-        if (
-            "actual_method" in form_data
-            and form_data["actual_method"] == "PUT"
-        ):
-          
-            knot_update = Knot.objects.get(pk=knot_id)
-            knot_update.name = form_data['name']
-            knot_update.rope_type = form_data['rope_type']
-            knot_update.company = form_data['company']
-            knot_update.price = form_data['price']
-            knot_update.notes = form_data['notes']
-            knot_update.video = form_data['video']
-            knot_update.image = form_data['image']
+            if (
+                "actual_method" in form_data
+                and form_data["actual_method"] == "PUT"
+            ):
+            
+                knot_update = Knot.objects.get(pk=knot_id)
+                knot_update.name = form_data['name']
+                knot_update.rope_type = form_data['rope_type']
+                knot_update.company = form_data['company']
+                knot_update.price = form_data['price']
+                knot_update.notes = form_data['notes']
+                knot_update.video = form_data['video']
+                knot_update.image = form_data['image']
 
-            knot_update.save()
+                knot_update.save()
 
-            return redirect(reverse('knotitapp:knots'))
+                return redirect(reverse('knotitapp:knots'))
 
-        if (
-            "actual_method" in form_data
-            and form_data["actual_method"] == "DELETE"
-           ):
-                
-            knot = Knot.objects.get(pk=knot_id)
-            knot.delete()
+            if (
+                "actual_method" in form_data
+                and form_data["actual_method"] == "DELETE"
+            ):
+                    
+                knot = Knot.objects.get(pk=knot_id)
+                knot.delete()
 
-            return redirect(reverse('knotitapp:knots'))
+                return redirect(reverse('knotitapp:knots'))
+    else:
+        return redirect(reverse('knotitapp:knots'))
